backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow
pydantic
python-dotenv
pyyaml
prometheus_client
